backend/dc_crawling/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
json2html==1.3.0
numpy==1.25.2
packaging==23.1
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions>=4.7.1,<5
tzdata==2023.3
urllib3==2.0.4
cryptography
selenium==4.31.0
python-dotenv